сменил тип с int на long во всех числовых свойствах модели ответа ММ

+++ DirectumRXAutoDeployer/Notifiers/Mattermost/Models/Response/Acknowledgement.cs
@@ -4,6 +4,6 @@
     {
         public string user_id { get; set; }
         public string post_id { get; set; }
-        public int acknowledged_at { get; set; }
+        public long acknowledged_at { get; set; }
     }
 }
